requirements.txt
streamlit>=1.30
pytest>=7.0
google-genai>=0.8
python-dotenv>=1.0
